modify the sequence table operation.
Currently, the sequence table stores two types of information:
1. The latest range ID for each table.
2. The auto-increment column values for tables that have such columns.

This table is managed by the `txservice::Sequence` class.
For DDL operations, it is necessary to update the corresponding range ID and auto-increment value (if the table has an auto-increment column).
Therefore, the operations on the sequence table are implemented as a sub-operation of `UpsertTableOp` and `UpsertTableIndexOp`.

The following DDL statements require updating the range ID in the sequence table:
`CREATE / DROP / TRUNCATE TABLE, ADD / DROP INDEX`

The following DDL statements require updating the auto-increment value (if the table has an auto-increment column):
`CREATE / DROP / TRUNCATE TABLE`