bug: fix cce.data_store_size error (#28)
* bug: fix cce.data_store_size error
If the sk ccmap is in dirty state, the sk generated by pk may be inserted to ccmap between the two requests of `ScanSliceDeltaSizeCc` and `DataSyncScanCc`. The `data_store_size_` of these cces are not updated and are still in unknown state. Just skip this cce in this round ckpt.
---------

Co-authored-by: yangsw <yangsw@chengzhang.com>